Split out generating NFT unlock proof to separate service. Use a service per NFT type. Currently only supporting ERC721. Added tests for ethers service.

+++ src/ownable/ownable.service.ts
@@ -1,16 +1,29 @@
 import { Injectable, OnModuleInit } from '@nestjs/common';
 import { PackageService } from '../package/package.service';
-import { EventChain } from '@ltonetwork/lto';
+import { Binary, EventChain, LTO } from '@ltonetwork/lto';
 import { ConfigService } from '../common/config/config.service';
 import { CosmWasmService } from '../cosmwasm/cosmwasm.service';
 import Contract from '../cosmwasm/contract';
 import fs from 'fs/promises';
+import { NFTInfo, OwnableInfo } from '../interfaces/OwnableInfo';
+import { integerToByteArray } from '@ltonetwork/lto/lib/utils/convert';
+import { NFTService } from '../nft/nft.service';
+
+interface InfoWithProof extends OwnableInfo {
+  proof: string;
+}
 
 @Injectable()
 export class OwnableService implements OnModuleInit {
   private readonly path: string;
 
-  constructor(private packages: PackageService, private config: ConfigService, private cosmWasm: CosmWasmService) {
+  constructor(
+    private packages: PackageService,
+    private config: ConfigService,
+    private cosmWasm: CosmWasmService,
+    private lto: LTO,
+    private nft: NFTService,
+  ) {
     this.path = this.config.get('chains.path');
   }
 
@@ -42,7 +55,7 @@ export class OwnableService implements OnModuleInit {
     }
   }
 
-  async accept(chain: EventChain): Promise<void> {
+  private async loadContract(chain: EventChain) {
     const packageCid: string = chain.events[0].parsedData.package;
 
     if (!(await this.packages.exists(packageCid))) {
@@ -56,18 +69,41 @@ export class OwnableService implements OnModuleInit {
 
     await this.apply(contract, chain);
 
+    return contract;
+  }
+
+  private verifyChainId(chain: EventChain, nft: NFTInfo): boolean {
+    const { publicKey, keyType } = chain.events[0].signKey as { publicKey: Binary; keyType: 'ed25519' | 'secp256k1' };
+    const account = this.lto.account({ publicKey: publicKey.base58, keyType });
+    const nonce = Binary.concat(
+      Binary.fromHex(nft.address),
+      nft.id.match(/^\d+$/) ? integerToByteArray(Number(nft.id)) : new Binary(nft.id),
+    );
+
+    const expectedId = new EventChain(account, nonce).id;
+
+    return chain.id === expectedId;
+  }
+
+  async accept(chain: EventChain): Promise<InfoWithProof> {
+    const contract = await this.loadContract(chain);
+
     const isLocked = await contract.query({ is_locked: {} });
     if (!isLocked) throw Error("Ownable isn't locked");
 
+    const info = (await contract.query({ get_info: {} })) as OwnableInfo;
+    if (!info.nft) {
+      throw new Error('Ownable is not associated with an NFT');
+    }
+
+    if (!this.verifyChainId(chain, info.nft)) {
+      throw new Error('Chain id mismatch: Unable to confirm the Ownable was forged for specified NFT');
+    }
+
     await fs.writeFile(`${this.path}/${chain.id}.json`, JSON.stringify(chain));
 
-    // Generate proof to unlock NFT
-    // The info contains the nft contract address and the token id.
-    // Get the challenge from the NFT contract for this token id.
-    // Sign the challenge to generate the proof.
+    const proof = await this.nft.getUnlockProof(info.nft);
 
-    // const info = await contract.query({ info: {} });
-    // const challenge = await nftContract.query({ challenge: { token_id: info.token_id } });
-    // this.ethers.signMessage(challenge);
+    return { ...info, proof };
   }
 }
